// src/interfaces/deposits.interface.ts
export interface DepositsInterface {
  deposit_id: string;
  date_of_deposit: string;
  bank_name: string;
  amount_deposited: string;
  account_number: string;
  receipt_id: string;
  description: string;
  depositor: string;
  imei: string;
  app_version: string;
  time_of_deposit: string;
  status: number;
  hub_id: string;
  comment: string;
  status_updated_date: string;
  reference_deposit_id: string;
  payment_details_id: string;
  image_id: string;
  payment_mode: string;
}